+++ testdata/test1.out
@@ -1,15 +1,12 @@
            2  4  4  2  1 
           ---------------
  1  1  1 | 1  0  1  0  1 
- 4       |    1  1  1    
- 2       | 0  1  1  0  0 
- 3       |    1  1     0 
- 1       |    1  0       
+ 4       | 1  1  1  1  0 
+ 2       | 0  1  1  1    
+ 3       | 0  1  1  1  0 
+ 1       | 0  1  0  0  0 
 ■□■□■
- ■■■ 
-□■■□□
- ■■ □
- ■□  
- 
- 
-□□□
+■■■■□
+□■■■ 
+□■■■□
+□■□□□
