+++ testdata/test6.out
@@ -4,33 +4,33 @@
                        1           2        1        1  2  4    
                        1           1                 1          
                    ---------------------------------------------
- 3  8  1          |    1  1  0     1  1  1  1  1  1  1          
- 1  5  1  1       |    0  0  0     0  1  1  1  1  1  0     0    
- 1  1  1  5  1    | 0  1  0  1  0  1  0  1  1  1  1  1  0  1    
+ 3  8  1          | 1  1  1  0  1  1  1  1  1  1  1  1  0  1  0 
+ 1  5  1  1       | 1  0  0  0  0  0  1  1  1  1  1  0  1  0  1 
+ 1  1  1  5  1    | 0  1  0  1  0  1  0  1  1  1  1  1  0  1  0 
  5  1             | 0  0  0  0  0  0  0  0  1  1  1  1  1  0  1 
- 2  4  3          |    1     0  0  1  1  1  1  0  0  0  1  1  1 
+ 2  4  3          | 1  1  0  0  0  1  1  1  1  0  0  0  1  1  1 
  3  4  2          | 1  1  1  0  1  1  1  1  0  0  0  0  0  1  1 
  7  1             | 1  1  1  1  1  1  1  0  0  0  0  0  0  0  1 
- 7  1  1  1       | 1  1  1  1  1  1  1  0  1  0           0  1 
+ 7  1  1  1       | 1  1  1  1  1  1  1  0  1  0  0  0  1  0  1 
  3  3  2  2  1    | 1  1  1  0  1  1  1  0  1  1  0  1  1  0  1 
- 3  1  1  1  1  1 | 1  1  1  0  1  0  1  0  0  1           0  1 
+ 3  1  1  1  1  1 | 1  1  1  0  1  0  1  0  0  1  0  0  1  0  1 
  4  2  2  2  1    | 1  1  1  1  0  1  1  0  1  1  0  1  1  0  1 
- 1  2  3  2       |    0     1  0  1  1  1  0  0  0  0  0  1  1 
- 1  1  2  1  2    |    1     1     0     1     0           1  1 
- 1  3  2  3       |    0     1  1        1     0           1  1 
- 1  3  4          |    1     1           0     0           1    
- ■■□ ■■■■■■■   
- □□□ □■■■■■□ □ 
-□■□■□■□■■■■■□■ 
+ 1  2  3  2       | 1  0  1  1  0  1  1  1  0  0  0  0  0  1  1 
+ 1  1  2  1  2    | 0  1  0  1  0  0  1  1  0  0  1  0  0  1  1 
+ 1  3  2  3       | 1  0  1  1  1  0  0  1  1  0  0  0  1  1  1 
+ 1  3  4          | 0  1  0  1  1  1  0  0  0  0  0  1  1  1  1 
+■■■□■■■■■■■■□■□
+■□□□□□■■■■■□■□■
+□■□■□■□■■■■■□■□
 □□□□□□□□■■■■■□■
- ■ □□■■■■□□□■■■
+■■□□□■■■■□□□■■■
 ■■■□■■■■□□□□□■■
 ■■■■■■■□□□□□□□■
-■■■■■■■□■□   □■
+■■■■■■■□■□□□■□■
 ■■■□■■■□■■□■■□■
-■■■□■□■□□■   □■
+■■■□■□■□□■□□■□■
 ■■■■□■■□■■□■■□■
- □ ■□■■■□□□□□■■
- ■ ■ □ ■ □   ■■
- □ ■■  ■ □   ■■
- ■ ■   □ □   ■ 
+■□■■□■■■□□□□□■■
+□■□■□□■■□□■□□■■
+■□■■■□□■■□□□■■■
+□■□■■■□□□□□■■■■
